backend/model_app/requirements.txt
pydantic-settings==2.13.1
uvicorn[standard]==0.32.1

torch==2.10.0
vllm==0.17.1
mistral-common==1.11.0

sentence-transformers==5.3.0
faiss-cpu==1.11.0.post1

motor==3.7.1
pymongo==4.16.0
redis==5.0.4

cachetools==5.3.3
numpy==2.2.2
pandas==2.2.2
pynvml==13.0.1
